docs: 프로젝트 페이즈 관리 시스템 구축
## 추가된 문서
- project_phases.md: 전체 페이즈별 진행 현황
- phase3_checklist.md: Phase 3 상세 체크리스트
- PROJECT_STATUS.md: 실시간 프로젝트 상태 대시보드

## 페이즈 구성
- Phase 1: 초기 설정 (✅ 완료)
- Phase 2: 환경 설정 및 데모 모드 (✅ 완료)
- Phase 3: 핵심 기능 구현 (🔄 시작)
- Phase 4-7: 계획됨

프로젝트 진행 상황을 체계적으로 관리할 수 있는 구조 마련

+++ lib/screens/weight_input_screen.dart
@@ -104,7 +104,7 @@ class _WeightInputScreenState extends ConsumerState<WeightInputScreen> {
   Widget build(BuildContext context) {
     final weight = double.tryParse(_weightController.text) ?? 0;
     final bmi = weight > 0 ? BMICalculator.calculateBMI(weight, userHeight) : 0;
-    final bmiCategory = weight > 0 ? BMICalculator.getBMICategory(bmi) : null;
+    final bmiCategory = weight > 0 ? BMICalculator.getBMICategory(bmi.toDouble()) : null;
     
     return Scaffold(
       backgroundColor: AppColors.background,
